refactor(P2.5 sub-slice C): move CTE join-condition rewriters into render_plan/cte_rewrite.rs
Phase-2 §5.1 module move, third sub-slice of the cte_rewrite group. Extracted the
contiguous join-condition CTE-rewrite group verbatim from plan_builder_utils.rs
into the existing render_plan/cte_rewrite.rs:
collect_with_cte_table_aliases / strip_table_alias_from_resolved /
rewrite_join_conditions_for_cte_aliases (old lines 1430-1627).

All three had zero external callers (the earlier grep "hits" in join_builder were
comments), so the pub(crate) use re-export serves only this module's internal
call sites. The only non-verbatim change: widened all three fn -> pub(crate) fn
so the re-export reaches them (documented at each site). super::-qualified deps
(CteSchemas, render_expr::*, UnionItems, FromTableItem, map_render_expr) resolve
from the new sibling module unchanged. Zero logic edits.

Gates: 1612 lib + 529 integration (incl. corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; byte-identical (moved block diff-clean vs main modulo the 3
visibility lines, no golden churn, ratchet net-zero); fmt/clippy clean;
warning-free --tests build.
