+++ src/render_plan/plan_builder_utils.rs
@@ -107,10 +107,12 @@ pub(crate) use super::plan_predicates::{has_with_clause_in_graph_rel, plan_conta
 // resolve during the transition. Sub-slice A: the expression-rewriting cluster
 // (`join_builder`'s external call + the two internal `_join` sites). Sub-slice B:
 // `remap_cte_names_in_render_plan` (2 internal sites; `remap_cte_names_in_expr`
-// has no caller left here so it is not re-exported).
+// has no caller left here so it is not re-exported). Sub-slice C: the
+// join-condition rewrite group (all called only internally in this module).
 pub(crate) use super::cte_rewrite::{
-    remap_cte_names_in_render_plan, rewrite_operator_application_for_cte,
-    rewrite_operator_application_for_cte_join,
+    collect_with_cte_table_aliases, remap_cte_names_in_render_plan,
+    rewrite_join_conditions_for_cte_aliases, rewrite_operator_application_for_cte,
+    rewrite_operator_application_for_cte_join, strip_table_alias_from_resolved,
 };
 
 /// Build property mapping from select items for CTE column resolution.
@@ -1425,205 +1427,6 @@ fn rewrite_count_to_conditional(agg: &mut AggregateFnCall, cond: RenderExpr) {
 /// variables that lack the `p{N}_{alias}_{prop}` column shape.
 pub(crate) fn quote_qualified_col(alias: &str, col: &str) -> String {
     format!("{}.{}", alias, current_function_mapper().quote_alias(col))
-}
-
-/// Collect all `with_*_cte_*` table aliases referenced in a RenderPlan's expressions.
-fn collect_with_cte_table_aliases(
-    plan: &crate::render_plan::RenderPlan,
-) -> std::collections::HashSet<String> {
-    use crate::render_plan::render_expr::RenderExpr;
-
-    fn collect_from_expr(expr: &RenderExpr, result: &mut std::collections::HashSet<String>) {
-        match expr {
-            RenderExpr::PropertyAccessExp(pa) => {
-                let alias = &pa.table_alias.0;
-                if alias.starts_with("with_") && alias.contains("_cte_") {
-                    result.insert(alias.clone());
-                }
-            }
-            RenderExpr::OperatorApplicationExp(op) => {
-                for operand in &op.operands {
-                    collect_from_expr(operand, result);
-                }
-            }
-            RenderExpr::AggregateFnCall(fc) => {
-                for arg in &fc.args {
-                    collect_from_expr(arg, result);
-                }
-            }
-            RenderExpr::ScalarFnCall(fc) => {
-                for arg in &fc.args {
-                    collect_from_expr(arg, result);
-                }
-            }
-            RenderExpr::Case(ce) => {
-                if let Some(ref expr) = ce.expr {
-                    collect_from_expr(expr, result);
-                }
-                for (cond, val) in &ce.when_then {
-                    collect_from_expr(cond, result);
-                    collect_from_expr(val, result);
-                }
-                if let Some(ref else_expr) = ce.else_expr {
-                    collect_from_expr(else_expr, result);
-                }
-            }
-            RenderExpr::ExistsSubquery(_) => {
-                // ExistsSubquery contains pre-rendered SQL, no expressions to scan
-            }
-            _ => {}
-        }
-    }
-
-    let mut result = std::collections::HashSet::new();
-    for item in &plan.select.items {
-        collect_from_expr(&item.expression, &mut result);
-    }
-    for join in &plan.joins.0 {
-        for op in &join.joining_on {
-            collect_from_expr(&RenderExpr::OperatorApplicationExp(op.clone()), &mut result);
-        }
-    }
-    if let Some(ref filter) = plan.filters.0 {
-        collect_from_expr(filter, &mut result);
-    }
-    for expr in &plan.group_by.0 {
-        collect_from_expr(expr, &mut result);
-    }
-    for item in &plan.order_by.0 {
-        collect_from_expr(&item.expression, &mut result);
-    }
-    result
-}
-
-/// Strip table alias from resolved CTE property accesses, recursively.
-/// Converts `PropertyAccessExp(table_alias, column)` → `Column(column)` so ORDER BY
-/// references output column aliases (visible after GROUP BY) instead of internal table references.
-fn strip_table_alias_from_resolved(expr: &RenderExpr) -> RenderExpr {
-    use super::render_expr::*;
-    use crate::graph_catalog::expression_parser::PropertyValue;
-    // Exhaustive combinator: drop the table alias off each resolved property
-    // access (PropertyAccess → bare Column), recurse structurally into every
-    // value-wrapper. The former hand-rolled walk handled
-    // PropertyAccess/Operator/ScalarFn/Aggregate and fell through
-    // `_ => expr.clone()` for List/Case/ArraySubscript/…, silently leaving a
-    // qualified column inside those wrappers (e.g. `ORDER BY [x.a][0]` over a
-    // CTE-scoped WITH). Latent (no corpus query reached it; byte-identical on
-    // migration), now structurally impossible.
-    map_render_expr(expr, &mut |node| match node {
-        RenderExpr::PropertyAccessExp(pa) => {
-            if let PropertyValue::Column(col) = &pa.column {
-                RenderRewrite::Replace(RenderExpr::Column(Column(PropertyValue::Column(
-                    col.clone(),
-                ))))
-            } else {
-                RenderRewrite::Replace(node.clone())
-            }
-        }
-        _ => RenderRewrite::Recurse,
-    })
-}
-
-/// Rewrite join conditions in a rendered plan that reference CTE aliases.
-/// When a join condition uses `cte_alias.base_column` (e.g., `friend.id`),
-/// replace it with the CTE's prefixed column (e.g., `friend.p6_friend_id`).
-fn rewrite_join_conditions_for_cte_aliases(
-    plan: &mut crate::render_plan::RenderPlan,
-    cte_references: &std::collections::HashMap<String, String>,
-    cte_schemas: &super::CteSchemas,
-) {
-    use crate::render_plan::render_expr::RenderExpr;
-
-    fn rewrite_expr_for_cte(
-        expr: RenderExpr,
-        cte_references: &std::collections::HashMap<String, String>,
-        cte_schemas: &super::CteSchemas,
-    ) -> RenderExpr {
-        match expr {
-            RenderExpr::PropertyAccessExp(mut pa) => {
-                let alias = &pa.table_alias.0;
-                if let Some(cte_name) = cte_references.get(alias) {
-                    if let Some(meta) = cte_schemas.get(cte_name) {
-                        let col_name = match &pa.column {
-                            crate::graph_catalog::expression_parser::PropertyValue::Column(c) => {
-                                c.clone()
-                            }
-                            crate::graph_catalog::expression_parser::PropertyValue::Expression(
-                                e,
-                            ) => e.clone(),
-                        };
-                        // Look up (alias, column) → CTE column name
-                        if let Some(cte_col) = meta
-                            .property_mapping
-                            .get(&(alias.clone(), col_name.clone()))
-                        {
-                            log::info!(
-                                "🔧 rewrite_join_cte: {}.{} → {}.{}",
-                                alias,
-                                col_name,
-                                alias,
-                                cte_col
-                            );
-                            pa.column =
-                                crate::graph_catalog::expression_parser::PropertyValue::Column(
-                                    cte_col.clone(),
-                                );
-                        }
-                    }
-                }
-                RenderExpr::PropertyAccessExp(pa)
-            }
-            RenderExpr::OperatorApplicationExp(mut op) => {
-                op.operands = op
-                    .operands
-                    .into_iter()
-                    .map(|o| rewrite_expr_for_cte(o, cte_references, cte_schemas))
-                    .collect();
-                RenderExpr::OperatorApplicationExp(op)
-            }
-            other => other,
-        }
-    }
-
-    for join in &mut plan.joins.0 {
-        for op in &mut join.joining_on {
-            let rewritten = rewrite_expr_for_cte(
-                RenderExpr::OperatorApplicationExp(op.clone()),
-                cte_references,
-                cte_schemas,
-            );
-            if let RenderExpr::OperatorApplicationExp(new_op) = rewritten {
-                *op = new_op;
-            }
-        }
-    }
-
-    // Also rewrite FROM table if it references a CTE
-    if let crate::render_plan::FromTableItem(Some(ref mut from_ref)) = plan.from {
-        if let Some(alias) = &from_ref.alias {
-            if let Some(cte_name) = cte_references.get(alias) {
-                if from_ref.name != *cte_name {
-                    log::info!(
-                        "🔧 rewrite_join_cte: Updating FROM table '{}' → '{}' for alias '{}'",
-                        from_ref.name,
-                        cte_name,
-                        alias
-                    );
-                    from_ref.name = cte_name.clone();
-                }
-            }
-        }
-    }
-
-    // CRITICAL: Also rewrite UNION branches' joins and FROM
-    // BidirectionalUnion creates UNION branches that are rendered independently.
-    // Their join conditions reference base table columns (e.g., person.id) which
-    // need CTE column rewriting just like the main plan's joins.
-    if let super::UnionItems(Some(ref mut union)) = plan.union {
-        for branch in &mut union.input {
-            rewrite_join_conditions_for_cte_aliases(branch, cte_references, cte_schemas);
-        }
-    }
 }
 
 /// Rewrite all occurrences of `old_alias` → `new_alias` in PropertyAccessExp table_alias
